% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connectDB}
\alias{connectDB}
\title{connectDB}
\usage{
connectDB(
  user = Sys.getenv("DB_USERNAME"),
  pass = Sys.getenv("DB_PASSWORD", rstudioapi::askForPassword())
)
}
\arguments{
\item{user}{user}

\item{pass}{pass}
}
\value{
Object of \code{\link{DBI::Connect}} with methods for ...
}
\description{
\code{connectDB} connects to db
}
\details{
This is a generic function:
}
\keyword{function}
