% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{dataframe}
\alias{dataframe}
\title{dataframe}
\usage{
dataframe()
}
\value{
Object of \code{\link{dataframe}} with revenue data
}
\description{
\code{path} returns a DataFrame with revenue data
}
\details{
details
}
\examples{
df<-path()
}
\keyword{function}
