% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{fetch}
\alias{fetch}
\title{fetch}
\usage{
fetch(sql, ...)
}
\arguments{
\item{sql}{sql statement}
}
\value{
Object of \code{\link{String}} with methods for ...
}
\description{
\code{fetch} returns a dataset
}
\details{
This is a generic function:
}
\examples{
result<-fetch("SELECT * FROM MY_TABLE")
}
\keyword{function}
