% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetchAll}
\alias{fetchAll}
\title{fetchAll}
\usage{
fetchAll()
}
\arguments{
\item{sql}{sql statement}
}
\value{
Object of \code{\link{String}} with methods for ...
}
\description{
\code{fetchAll} returns a dataset
}
\details{
This is a generic function:
}
\examples{
result<-fetchAll()
}
\keyword{function}
