% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installer.R
\name{installer}
\alias{installer}
\title{installer}
\usage{
installer(pkg)
}
\description{
\code{installer} returns a Dataframe with key data
}
\details{
details ...
}
\examples{
installer(packages)
}
\keyword{function}
