% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{myfunc}
\alias{myfunc}
\title{Test function}
\usage{
myfunc(a)
}
\arguments{
\item{a}{a}
}
\value{
Object of \code{\link{function}} with methods for ...
}
\description{
`myfunc` returns a function
}
\details{
details
}
\examples{
result<-function(2)
}
\keyword{function}
