% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{myfunc2}
\alias{myfunc2}
\title{Test function2}
\usage{
myfunc2(W)
}
\arguments{
\item{a}{a}
}
\value{
Object of \code{\link{function}} with methods for ...
}
\description{
\code{myfunc2} returns a function
}
\details{
details
}
\examples{
result<-function(2)
}
\keyword{function}
