% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installer.R
\docType{data}
\name{packages.needed}
\alias{packages.needed}
\title{installer}
\format{
An object of class \code{character} of length 5.
}
\usage{
packages.needed
}
\description{
\code{installer} returns a Dataframe with key data
}
\details{
details ...
}
\examples{
installer(packages)
}
\keyword{function}
