% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{path}
\alias{path}
\title{path}
\usage{
path()
}
\value{
Object of \code{\link{dataframe}} with revenue data
}
\description{
\code{path} returns a DataFrame with revenue data
}
\details{
details
}
\examples{
df<-path()
}
\keyword{function}
