% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\docType{data}
\name{pkg.db.env}
\alias{pkg.db.env}
\title{connectDB}
\format{
An object of class \code{environment} of length 3.
}
\usage{
pkg.db.env
}
\arguments{
\item{user}{user}

\item{pass}{pass}
}
\value{
Object of \code{\link{DBI::Connect}} with methods for ...
}
\description{
\code{connectDB} connects to db
}
\details{
This is a generic function:
}
\examples{
result<-connectDB()
}
\keyword{function}
