% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{revenue}
\alias{revenue}
\title{revenue
Revenue data}
\format{
A table with x rows and y variables
\describe{
\item{price}{price in US dollars (\$326--\$18,823)}
\item{carat}{weight of the diamond (0.2--5.01)}
\item{cut}{quality of the cut (Fair, Good, Very Good, Premium, Ideal)}
}
}
\source{
\url{http://www.fom.de/}
}
\usage{
data(revenue)
}
\description{
A dataset containing the revenue and other attributes of x companies.
}
\examples{
View(data(revenue))
}
\keyword{datasets}
