% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\title{summary}
\usage{
summary()
}
\value{
Object of \code{\link{dataframe}} with revenue data
}
\description{
\code{path} returns a Dataframe with key data
}
\details{
details ...
}
\examples{
df<-summary()
}
\keyword{function}
