% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{Dataframe}
\alias{Dataframe}
\title{Dataframe}
\usage{
Dataframe()
}
\value{
Object of \code{\link{dataframe}} with revenue data
}
\description{
\code{Dataframe} returns a DataFrame with revenue data
}
\details{
details
}
\examples{
df<-Dataframe()
}
\keyword{function}
