% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{Test function1}
\usage{
test(a, b)
}
\arguments{
\item{a}{a,a}

\item{b}{b}
}
\value{
Object of \code{\link{String}} with methods for ...
}
\description{
`test` returns test
}
\details{
This is a generic function: methods can be defined for it directly
or via the [Summary()] group generic. For this to work properly,
the arguments `...` should be unnamed, and dispatch is on the
first argument.
}
\examples{
result<-test("a",b)
}
\keyword{function}
