% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{Constructor}
\usage{
test(a, b)
}
\arguments{
\item{provider}{a,b}
}
\value{
Object of \code{\link{String}} with methods for ...
}
\description{
Constructor
}
\examples{
result<-test(a,b)
}
\keyword{function}
